Moved controllers into actual controllers
