--- src/http/api/derepost.ts
@@ -1,144 +1,0 @@
-import { createHash } from 'node:crypto';
-
-import { type Actor, Announce, PUBLIC_COLLECTION, Undo } from '@fedify/fedify';
-import { type AppContext, globalFedify } from 'app';
-import { exhaustiveCheck, getError, getValue, isError } from 'core/result';
-import { parseURL } from 'core/url';
-import { lookupActor, lookupObject } from 'lookup-helpers';
-import type { KnexPostRepository } from 'post/post.repository.knex';
-import type { PostService } from 'post/post.service';
-import { ACTOR_DEFAULT_HANDLE } from '../../constants';
-
-export function createDerepostActionHandler(
-    postService: PostService,
-    postRepository: KnexPostRepository,
-) {
-    return async function derepostAction(ctx: AppContext) {
-        const account = ctx.get('account');
-        const id = ctx.req.param('id');
-        const apCtx = globalFedify.createContext(ctx.req.raw as Request, {
-            globaldb: ctx.get('globaldb'),
-            logger: ctx.get('logger'),
-        });
-
-        const post = await lookupObject(apCtx, id);
-        if (!post) {
-            return new Response(JSON.stringify({ error: 'Post not found' }), {
-                status: 404,
-            });
-        }
-
-        const announceId = apCtx.getObjectUri(Announce, {
-            id: createHash('sha256').update(post.id!.href).digest('hex'),
-        });
-
-        const undoId = apCtx.getObjectUri(Undo, {
-            id: createHash('sha256').update(announceId.href).digest('hex'),
-        });
-
-        const announceToUndoJson = await ctx
-            .get('globaldb')
-            .get([announceId.href]);
-
-        if (!announceToUndoJson) {
-            return new Response(
-                JSON.stringify({ error: 'Repost activity not found' }),
-                {
-                    status: 409,
-                },
-            );
-        }
-
-        const announceToUndo = await Announce.fromJsonLd(announceToUndoJson);
-
-        const actor = await apCtx.getActor(ACTOR_DEFAULT_HANDLE); // TODO This should be the actor making the request
-
-        const idAsUrl = parseURL(id);
-
-        if (!idAsUrl) {
-            return new Response(
-                JSON.stringify({ error: 'ID should be a valid URL' }),
-                {
-                    status: 400,
-                },
-            );
-        }
-
-        const originalPostResult = await postService.getByApId(idAsUrl);
-
-        if (isError(originalPostResult)) {
-            const error = getError(originalPostResult);
-            switch (error) {
-                case 'upstream-error':
-                    ctx.get('logger').info(
-                        'Upstream error fetching post for dereposting',
-                        { postId: idAsUrl.href },
-                    );
-                    break;
-                case 'not-a-post':
-                    ctx.get('logger').info(
-                        'Resource for dereposting is not a post',
-                        { postId: idAsUrl.href },
-                    );
-                    break;
-                case 'missing-author':
-                    ctx.get('logger').info(
-                        'Post for dereposting has missing author',
-                        { postId: idAsUrl.href },
-                    );
-                    break;
-                default:
-                    return exhaustiveCheck(error);
-            }
-        } else {
-            const originalPost = getValue(originalPostResult);
-            originalPost.removeRepost(account);
-            await postRepository.save(originalPost);
-        }
-
-        const undo = new Undo({
-            id: undoId,
-            actor: actor,
-            object: announceToUndo,
-            to: PUBLIC_COLLECTION,
-            cc: apCtx.getFollowersUri(ACTOR_DEFAULT_HANDLE),
-        });
-
-        // Add the undo activity to the database
-        const undoJson = await undo.toJsonLd();
-        await ctx.get('globaldb').set([undo.id!.href], undoJson);
-
-        // Remove announce activity from database
-        await ctx.get('globaldb').delete([announceId.href]);
-
-        // Send the undo activity
-        let attributionActor: Actor | null = null;
-        if (post.attributionId) {
-            attributionActor = await lookupActor(
-                apCtx,
-                post.attributionId.href,
-            );
-        }
-        if (attributionActor) {
-            apCtx.sendActivity(
-                { username: account.username },
-                attributionActor,
-                undo,
-                {
-                    preferSharedInbox: true,
-                },
-            );
-        }
-
-        apCtx.sendActivity({ username: account.username }, 'followers', undo, {
-            preferSharedInbox: true,
-        });
-
-        return new Response(JSON.stringify(undoJson), {
-            headers: {
-                'Content-Type': 'application/activity+json',
-            },
-            status: 200,
-        });
-    };
-}
