+++ src/http/api/feed.controller.ts
@@ -1,10 +1,9 @@
 import * as Sentry from '@sentry/node';
-
+import type { AccountService } from 'account/account.service';
+import { getAccountHandle } from 'account/utils';
+import type { AppContext } from 'app';
+import type { FeedService, FeedType } from 'feed/feed.service';
 import type { PostInteractionCountsService } from 'post/post-interaction-counts.service';
-import type { AccountService } from '../../account/account.service';
-import { getAccountHandle } from '../../account/utils';
-import type { AppContext } from '../../app';
-import type { FeedService, FeedType } from '../../feed/feed.service';
 import type { PostDTO } from './types';
 
 /**
@@ -17,25 +16,58 @@ const DEFAULT_FEED_POSTS_LIMIT = 20;
  */
 const MAX_FEED_POSTS_LIMIT = 100;
 
-/**
- * Create a handler to handle a request for a user's feed
- *
- * @param feedService Feed service instance
- * @param accountService Account service instance
- * @param feedType Type of feed
- */
-export function createGetFeedHandler(
-    feedService: FeedService,
-    accountService: AccountService,
-    postInteractionCountsService: PostInteractionCountsService,
-    feedType: FeedType,
-) {
+export class FeedController {
+    constructor(
+        private readonly feedService: FeedService,
+        private readonly accountService: AccountService,
+        private readonly postInteractionCountsService: PostInteractionCountsService,
+    ) {}
+
     /**
-     * Handle a request for a user's feed
+     * Handle a request for the main feed
      *
      * @param ctx App context
      */
-    return async function handleGetFeed(ctx: AppContext) {
+    async handleGetFeed(ctx: AppContext) {
+        return this.handleGetFeedByType(ctx, 'Feed');
+    }
+
+    /**
+     * Handle a request for the inbox
+     *
+     * @param ctx App context
+     */
+    async handleGetInbox(ctx: AppContext) {
+        return this.handleGetFeedByType(ctx, 'Inbox');
+    }
+
+    /**
+     * Handle a request for the notes feed
+     *
+     * @param ctx App context
+     */
+    async handleGetNotesFeed(ctx: AppContext) {
+        // Same as handleGetFeed but could be customized in the future
+        return this.handleGetFeedByType(ctx, 'Feed');
+    }
+
+    /**
+     * Handle a request for the reader feed
+     *
+     * @param ctx App context
+     */
+    async handleGetReaderFeed(ctx: AppContext) {
+        // Same as handleGetInbox but could be customized in the future
+        return this.handleGetFeedByType(ctx, 'Inbox');
+    }
+
+    /**
+     * Handle a request for a user's feed by type
+     *
+     * @param ctx App context
+     * @param feedType Type of feed
+     */
+    private async handleGetFeedByType(ctx: AppContext, feedType: FeedType) {
         const queryCursor = ctx.req.query('next');
         const cursor = queryCursor ? decodeURIComponent(queryCursor) : null;
 
@@ -52,7 +84,7 @@ export function createGetFeedHandler(
 
         const account = ctx.get('account');
 
-        const { results, nextCursor } = await feedService.getFeedData({
+        const { results, nextCursor } = await this.feedService.getFeedData({
             accountId: account.id,
             feedType,
             limit,
@@ -119,7 +151,7 @@ export function createGetFeedHandler(
         // Request an update of the interaction counts for the posts in the
         // feed - We do not await this as we do not want to increase the
         // response time of the request
-        postInteractionCountsService
+        this.postInteractionCountsService
             .requestUpdate(
                 ctx.get('site').host,
                 results.map((post) => post.post_id),
@@ -142,5 +174,5 @@ export function createGetFeedHandler(
                 status: 200,
             },
         );
-    };
+    }
 }
