--- src/http/api/index.ts
@@ -1,8 +1,0 @@
-export * from './account';
-export * from './feed';
-export * from './notification';
-export * from './note';
-export * from './post';
-export * from './search';
-export * from './image-upload';
-export * from './webhook';
