+++ src/http/api/media.controller.ts
@@ -1,13 +1,20 @@
 import type { AccountService } from 'account/account.service';
+import type { AppContext } from 'app';
 import { exhaustiveCheck, getError, getValue, isError } from 'core/result';
-import type { Context } from 'hono';
 import type { ImageStorageService } from 'storage/image-storage.service';
 
-export function createImageUploadHandler(
-    accountService: AccountService,
-    imageStorageService: ImageStorageService,
-) {
-    return async function handleImageUpload(ctx: Context) {
+export class MediaController {
+    constructor(
+        private readonly accountService: AccountService,
+        private readonly imageStorageService: ImageStorageService,
+    ) {}
+
+    /**
+     * Handle an image upload request
+     *
+     * @param ctx App context
+     */
+    async handleImageUpload(ctx: AppContext) {
         const logger = ctx.get('logger');
         const formData = await ctx.req.formData();
         const file = formData.get('file');
@@ -16,8 +23,10 @@ export function createImageUploadHandler(
             return new Response('No valid file provided', { status: 400 });
         }
 
-        const account = await accountService.getAccountForSite(ctx.get('site'));
-        const result = await imageStorageService.save(
+        const account = await this.accountService.getAccountForSite(
+            ctx.get('site'),
+        );
+        const result = await this.imageStorageService.save(
             file,
             `images/${account.uuid}/`,
         );
@@ -51,5 +60,5 @@ export function createImageUploadHandler(
             },
             status: 200,
         });
-    };
+    }
 }
