+++ src/http/api/webhook.controller.ts
@@ -1,8 +1,7 @@
-import { z } from 'zod';
-
+import type { AppContext } from 'app';
 import { exhaustiveCheck, getError, getValue, isError } from 'core/result';
 import type { PostService } from 'post/post.service';
-import type { AppContext } from '../../app';
+import { z } from 'zod';
 import { postToDTO } from './helpers/post';
 import { BadRequest } from './helpers/response';
 
@@ -35,13 +34,15 @@ const PostPublishedWebhookSchema = z.object({
     }),
 });
 
-/**
- * Handle a post.published webhook
- *
- * @param ctx App context instance
- */
-export function createPostPublishedWebhookHandler(postService: PostService) {
-    return async function handleWebhookPostPublished(ctx: AppContext) {
+export class WebhookController {
+    constructor(private readonly postService: PostService) {}
+
+    /**
+     * Handle a post.published webhook
+     *
+     * @param ctx App context instance
+     */
+    async handlePostPublished(ctx: AppContext) {
         let data: PostInput;
 
         try {
@@ -57,7 +58,7 @@ export function createPostPublishedWebhookHandler(postService: PostService) {
 
         const account = ctx.get('account');
 
-        const postResult = await postService.handleIncomingGhostPost(
+        const postResult = await this.postService.handleIncomingGhostPost(
             account,
             data,
         );
@@ -90,5 +91,5 @@ export function createPostPublishedWebhookHandler(postService: PostService) {
             },
             status: 200,
         });
-    };
+    }
 }
